refactor: extract shared useCountUp hook
